requirements.txt
# Telegram Bot
python-telegram-bot>=21.0

# Gemini API
google-generativeai>=0.8.0

# Web Crawling
httpx>=0.27.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Database
aiosqlite>=0.20.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
